feat: role-based capabilities, smart mode detection, payment recording, subadmin access
- Toshi now uses a capability map instead of hardcoded role gates
- Smart mode detection: auto-switches setup→assistant when user asks a question
- Fee payment recording: migration, model, service method, admin UI, Toshi flow
- SiteSubadmin (usergroup_id 2) now has Toshi access and middleware support
- keyword route for create school uses regex for natural language matching
- LLM prompt includes allowed actions and 'create school' trigger instruction
- Accountant gets record_payment capability

Fixes: schooladmin middleware handles usergroup_id 2 redirect
Fixes: buildContext treats usergroup_id 2 as platform user

+++ database/migrations/2026_06_23_074056_create_whatsapp_pending_parent_links_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->id();
             $table->string('phone', 20);                          // Parent's WhatsApp number
             $table->unsignedBigInteger('school_id');               // School from matched student
-            $table->unsignedBigInteger('student_id')->nullable();  // Student user ID
+            $table->unsignedInteger('student_id')->nullable();   // Student user ID (matches users.id type)
             $table->string('student_payment_code', 50);            // The code they entered
             $table->string('student_name')->nullable();            // For display in admin panel
             $table->string('status', 20)->default('pending');      // pending | approved | rejected
